Transformers.js line items shouldn't load resources in the benchmark
Right now some of the resources are loaded via relative paths in the
benchmark. Instead of trying to patch the benchmark to use the blob,
this patch makes it easier to redirect `fetch` so that it can find
the already loaded data.

I did this by adding a `JetStream.resources` object that maps from
the original relative path resource to the blobURLOrPath vended
by the preloader.