Errors with Non-Standard Query Params Should Be Red
Right now we rotate all colors in the body so the current the red color
ends up green. We should rotate the error class back so that it ends up
red again. I did this via a variable so if we change the rotate in the
future error will automatically fix itself.